update : 회원가입 UX 개선
